fix: handle float annotation indices from Google Chat API
Google Chat sends annotation startIndex/length as floats (e.g. 0.0)
which cannot be unmarshaled into Go int fields. Changed to float64.

Also set default logging level to debug in install.sh for easier
troubleshooting during initial deployments.

+++ extras/scion-chat-app/internal/googlechat/adapter.go
@@ -398,9 +398,9 @@ type rawThread struct {
 }
 
 type rawAnnotation struct {
-	Type       string `json:"type"`
-	StartIndex int    `json:"startIndex"`
-	Length     int    `json:"length"`
+	Type       string  `json:"type"`
+	StartIndex float64 `json:"startIndex"`
+	Length     float64 `json:"length"`
 }
 
 type rawUser struct {
